Move search page to app router
